+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next
 
+- **[Fix]** Mark `deploy` property of `LibTarget.typedoc` as optional.
 - **[Internal]** Update self-dependency to `v0.15.0-beta.10`.
 
 ## 0.15.0-beta.10 (2017-08-28)